+++ src/App.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, {Component} from 'react';
 
 import './App.css';
 import HomeView from "./components/views/HomeView";
@@ -7,7 +7,7 @@ class App extends Component {
   render() {
     return (
       <div>
-          <HomeView/>
+        <HomeView/>
       </div>
 
     );
