+++ src/components/views/SearchResultsGridView.js
